+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based now on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Allow to upload files from a `Tempfile` object
+- Allow to upload files when mime-type can't be determined
+
 ## [2.1.2] - 2018-06-11
 
 ### Fixed